+++ docs/ReleaseNotes.rst
@@ -200,6 +200,12 @@ identified.  The improvements since the 3.8 release include:
 
   Replaces creation of ``std::shared_ptr`` from new expression with call to ``std::make_shared``.
 
+- New `modernize-operator-void-pointer
+  <http://clang.llvm.org/extra/clang-tidy/checks/modernize-operator-void-pointer.html>`_ check
+
+  Finds ``operator const void*`` overloads. These should often be
+  ``explicit operator bool`` overloads.
+
 - New `modernize-raw-string-literal
   <http://clang.llvm.org/extra/clang-tidy/checks/modernize-raw-string-literal.html>`_ check